# cinnamon-settings: noticeably accelerated startup time