This pull request boosts the startup time of cinnamon-settings noticeably by making the following changes:

1. ~Compile all python files in the `cinnamon-common` package during installation by using the python3 debhelper~ Edit: dropped as discussed below
2. Defer loading of heavier python packages (like `requests` or `imtools`, both ~50ms on my AMD Zen 3 dev machine). This has a noticeable effect, because all loaded python modules (`cs_*.py`) and their dependencies will be loaded at start time. Instead, import them before they are really used. Thus, the startup time will be shorter in any case.
3. The largest change attacks the before mentioned behavior that **all** modules are loaded in *any case*. This means that the "import (time) heavy" settings modules like the cinnamon spices x-lets are also load, if settings pages are started as standalone apps (i.e. using the menu shortcuts).

   The last commit therefore changes the way the the main `cinnamon-settings` is initialized. It now directly distinguishes between **starting into the "system settings"**, i.e. giving the overview over all settings modules, and the **standalone-startup of a given module** via the `argv`. Lazy-loading CCC and the real standalone modules (i.e. external binaries) are loaded in both cases as it made ~5ms difference while being more complex to handle.

As a result, I can see huge gains in startup performance. I tested both on my powerful development machine and on my Raspberry Pi 4. I didn't know how to reliably measure the time from start to visible GUI, so I just took the bash `time` internal and compared the CPU user and sys times. Here are some values (each one best of three measurements):

### Before

|             Command           |    Dev PC    |         RPi 4          | 
|----------------------------------|-----------------|-----------------------|
| `cinnamon-settings`                     | `user 0m0,350s`<br>`sys 0m0,058s` | `user 0m1,942s`<br>`sys 0,273ms` |
| `cinnamon-settings background` | `user 0m0,431s`<br>`sys 0m0,062s` | `user 0m2,203s`<br>`sys 0,352ms` |
| `cinnamon-settings user`             | `user 0m0,395s`<br>`sys 0m0,056s` | `user 0m1,959s`<br>`sys 0,325ms` |

### After

|             Command           |    Dev PC    |         RPi 4          | 
|----------------------------------|-----------------|-----------------------|
| `cinnamon-settings`                    | `user 0m0,206s`<br>`sys 0m0,027s` | `user 0m1,252s`<br>`sys 0m0,212s` |
| `cinnamon-settings background` | `user 0m0,258s`<br>`sys 0m0,035s` | `user 0m1,415s`<br>`sys 0m0,254s` |
| `cinnamon-settings user`            | `user 0m0,179s`<br>`sys 0m0,020s` | `user 0m1,053s`<br>`sys 0m0,203s` |

The commits contain also smaller other fixes and improvements to the existing code, hope that's okay for you.