@Fantu 2022-04-10
@okaestne thanks for the improvements
about dh-python was always keeped in debian package and I readded here in muffin rebase branch when I did some packaging improvements (as I saw it was removed): https://github.com/linuxmint/cinnamon/commit/6c93a890b3701934615f1d0dc68b156d08c9c7c9
about "--no-ext-rename" (in dh_python3) is because you spotted an issue without it? and should not be the path (usr/share/cinnamon) absolute instead relative? the compile is not done in package build but in package install (is in postinst/postrm)