@okaestne 2022-04-10
@Fantu thanks for your quick feedback. Should I drop the PyCompile commit in favor of your commit? The `--no-ext-rename` was just blindly taken from some other package tbh. It definitely works like this, also with the path.