@Fantu 2022-04-10
there aren't library files in the path or other that can cause unexpected cases where "--no-ext-rename" is useful (FWIK) and there was not issue spotted or reported by users in debian package related to python compiled files
maybe it is better to keep the one I had put back which is more tested and same as the debian packages (so there are also fewer differences)